fix import issue: move ConnectionTypes definition in io.sonata_config
